Long id!

+++ Examples/coursej/src/main/java/ru/mirea/oop/practice/coursej/api/vk/DocumentsApi.java
@@ -15,7 +15,7 @@ public interface DocumentsApi extends ExternalCall {
 
     Document[] list(Integer count,
                     Integer offset,
-                    Integer idOwner) throws IOException;
+                    Long idOwner) throws IOException;
 
     Document[] id(String docs) throws IOException;
 }
